added companion query

+++ src/main/scala/selection/selector/RTreeSelector.scala
@@ -14,7 +14,7 @@ class RTreeSelector(override val queryRange: Rectangle,
    override def query[T <: Shape : ClassTag](dataRDD: RDD[(Int, T)]): RDD[(Int, T)] = {
     val c = capacity.getOrElse({
       val dataSize = dataRDD.count
-      sqrt(dataSize / dataRDD.getNumPartitions)
+      sqrt(dataSize / dataRDD.getNumPartitions).toInt
     }) // rule for capacity calculation if not given
     val dataRDDWithIndex = dataRDD
       .map(x => x._2)
